`Refactor` modulo de usuarios
