+++ src/app/modules/herramientas/guards/ToolsGuard.guard.ts
@@ -2,7 +2,7 @@ import { inject } from "@angular/core";
 import { Router } from "@angular/router";
 
 import { AuthService } from "@modules/auth/services/auth.service";
-import { Nom_Rol } from "@modules/shared/interfaces";
+import { Nom_Rol } from "@modules/usuarios/interfaces";
 
 const forbiddenRoles = [Nom_Rol.INVITADO];
