+++ src/app/modules/shared/interfaces/index.ts
@@ -1,4 +1,1 @@
-export * from './Idu_Rol.enum';
-export * from './Nom_Rol.enum';
 export * from './checkmarx.interface';
-export * from './usuario.interface';
