+++ src/app/modules/usuarios/interfaces/iduRol.enum.ts
