+++ src/app/modules/usuarios/interfaces/index.ts
@@ -1,2 +1,6 @@
-export * from './OriginMethod.interface';
-export * from './UsersData.interface';
+export * from './iduRol.enum';
+export * from './nomRol.enum';
+export * from './originMethod.interface';
+export * from './usersData.interface';
+export * from './usuario.interface';
+export * from './usuarioPosition.interface';
