+++ src/app/modules/usuarios/interfaces/usuarioPosition.interface.ts
@@ -1,0 +1,7 @@
+import { Idu_Rol } from "./iduRol.enum";
+import { Nom_Rol } from "./nomRol.enum";
+
+export interface Position {
+    idu_rol: Idu_Rol;
+    nom_rol: Nom_Rol
+}
