+++ src/app/modules/usuarios/services/usuarios.service.ts
@@ -2,11 +2,10 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { BehaviorSubject, catchError, delay, map, Observable, of, switchMap, tap, throwError } from 'rxjs';
 
-import { Usuario } from '@modules/shared/interfaces/usuario.interface';
 import { environment } from '../../../../environments/environment';
 import { dataPerPage } from '@modules/shared/helpers/dataPerPage';
 import { NotificationsService } from '@modules/shared/services/notifications.service';
-import { OriginMethod, UsersData } from '../interfaces';
+import { OriginMethod, UsersData, Usuario } from '../interfaces';
 
 @Injectable({
   providedIn: 'root'
